requirements.txt
python-telegram-bot
websocket-client
python-dotenv
requests
aiohttp